travis.yml: explicitly ignore flake8 rules
When using the '--ignore' flag to flake8 on the command-line, ignore
rules specified in the .flake8 config are ignored. We therefore need to
explicitly list W503 and W504 again when invoking flake8 on
files/folders for which docstrings don't pass the test yet.

This commit also re-enables flake8 to run on all source files again, not
just those which were modified from master.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>