Replace 'diag' backend function
The 'from_vector_to_diagonal_matrix' function (previously in the
'SymmetricMatrices' class) has the same behavior as the current 'diag'
implementation present in the backends. For clarity, this change removes
the 'diag' function in favor of the 'from_vector_to_diagonal_matrix'.
The function lives in the new 'geomstats.utils' submodule as its
implementation is backend-agnostic.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>